requirements.txt
gradio
discord
cohere
python-dotenv